renames backend imports to common/api
